optimize get additional accounts
