+++ packages/blast-utilities/account-utils.js
@@ -3,44 +3,8 @@ const path = require('path')
 const bip39 = require('bip39')
 const { DirectSecp256k1HdWallet } = require('cudosjs')
 
-const { executeNodeMultiCmd } = require('./run-docker-commands')
-const { transferTokensByNameCommand } = require('./blast-helper')
 const { getProjectRootPath } = require('./package-info')
 const BlastError = require('./blast-error')
-const {
-  getAdditionalAccounts,
-  getAdditionalAccountsBalances,
-  getAddressPrefix
-} = require('./config-utils')
-
-async function createLocalAccountsFile() {
-  let localAccounts = require('../blast-config/default-accounts.json')
-  const numberOfAdditionalAccounts = getAdditionalAccounts()
-  if (numberOfAdditionalAccounts > 0) {
-    const additionalAccounts = {}
-    const customBalance = getAdditionalAccountsBalances()
-    const addressPrefix = getAddressPrefix()
-    for (let i = 1; i <= numberOfAdditionalAccounts; i++) {
-      const mnemonic = bip39.generateMnemonic(256)
-      const address = await getAddressFromMnemonic(mnemonic, addressPrefix)
-      // save the generated additional account so all accounts can be saved to a file later
-      additionalAccounts[`account${10 + i}`] = {
-        mnemonic: mnemonic,
-        address: address
-      }
-      // add new account from mnemonic to the local node and fund it
-      executeNodeMultiCmd(
-        `echo ${mnemonic} | cudos-noded keys add account${10 + i} --recover --keyring-backend test && ` +
-        transferTokensByNameCommand('faucet', `account${10 + i}`, `${customBalance}`)
-      )
-    }
-    localAccounts = {
-      ...localAccounts,
-      ...additionalAccounts
-    }
-  }
-  saveAccounts(localAccounts)
-}
 
 function getAccounts() {
   const configPath = path.join(getProjectRootPath(), 'accounts.json')
@@ -54,13 +18,22 @@ function getPrivateAccounts() {
   return privateAccounts
 }
 
+async function generateRandomAccount(addressPrefix) {
+  const mnemonic = bip39.generateMnemonic(256)
+  const address = await getAddressFromMnemonic(mnemonic, addressPrefix)
+  return {
+    mnemonic: mnemonic,
+    address: address
+  }
+}
+
 async function getAddressFromMnemonic(mnemonic, addressPrefix) {
   const wallet = await DirectSecp256k1HdWallet.fromMnemonic(mnemonic, { prefix: addressPrefix })
   const [acc] = await wallet.getAccounts()
   return acc.address
 }
 
-function saveAccounts(accounts) {
+function createLocalAccountsFile(accounts) {
   const accountFilePath = path.join(getProjectRootPath(), 'accounts.json')
   // delete accounts file if exists
   fs.rmSync(accountFilePath, { force: true })
@@ -73,7 +46,8 @@ function saveAccounts(accounts) {
 }
 
 module.exports = {
-  createLocalAccountsFile: createLocalAccountsFile,
   getAccounts: getAccounts,
-  getPrivateAccounts: getPrivateAccounts
+  getPrivateAccounts: getPrivateAccounts,
+  generateRandomAccount: generateRandomAccount,
+  createLocalAccountsFile: createLocalAccountsFile
 }
